+++ src/sodaspark/scan.py
@@ -6,13 +6,17 @@ from types import TracebackType
 from typing import Any
 
 from pyspark.sql import DataFrame, Row, SparkSession
+from pyspark.sql import types as T  # noqa: N812
 from sodasql.common.yaml_helper import YamlHelper
 from sodasql.dialects.spark_dialect import SparkDialect
 from sodasql.scan.file_system import FileSystemSingleton
+from sodasql.scan.measurement import Measurement
 from sodasql.scan.scan import Scan
+from sodasql.scan.scan_error import ScanError
 from sodasql.scan.scan_result import ScanResult
 from sodasql.scan.scan_yml import ScanYml
 from sodasql.scan.scan_yml_parser import ScanYmlParser
+from sodasql.scan.test_result import TestResult
 from sodasql.scan.warehouse import Warehouse
 from sodasql.scan.warehouse_yml import WarehouseYml
 from sodasql.soda_server_client.soda_server_client import SodaServerClient
@@ -48,7 +52,7 @@ class Cursor:
         self,
     ) -> list[tuple[str, str, None, None, None, None, bool]]:
         """
-        The description.
+        Get the description.
 
         Returns
         -------
@@ -199,7 +203,7 @@ class _SparkDialect(SparkDialect):
 
 def create_scan_yml(scan_definition: str | Path) -> ScanYml:
     """
-    Create a scan yml
+    Create a scan yml.
 
     Parameters
     ----------
@@ -249,7 +253,7 @@ def create_scan(
     scan_yml: ScanYml, soda_server_client: SodaServerClient | None = None
 ) -> Scan:
     """
-    Create a scan object
+    Create a scan object.
 
     Parameters
     ----------
@@ -273,11 +277,150 @@ def create_scan(
     return scan
 
 
+def measurements_to_data_frame(measurements: list[Measurement]) -> DataFrame:
+    """
+    Convert measurements to a data frame.
+
+    Parameters
+    ----------
+    measurements: List[Measurement]
+        The measurements.
+
+    Returns
+    -------
+    out : DataFrame
+        The measurements as data frame.
+    """
+    schema_group_values = T.StructType(
+        [
+            T.StructField(
+                "group", T.MapType(T.StringType(), T.StringType()), True
+            ),
+            T.StructField("value", T.StringType(), True),
+        ]
+    )
+    schema = T.StructType(
+        [
+            T.StructField("metric", T.StringType(), True),
+            T.StructField("column_name", T.StringType(), True),
+            T.StructField("value", T.StringType(), True),
+            T.StructField(
+                "group_values", T.ArrayType(schema_group_values, True), True
+            ),
+        ]
+    )
+    spark_session = SparkSession.builder.getOrCreate()
+    out = spark_session.createDataFrame(
+        measurements,
+        schema=schema,
+    )
+    return out
+
+
+def test_results_to_data_frame(test_results: list[TestResult]) -> DataFrame:
+    """
+    Convert TestResults to a data frame.
+
+    Parameters
+    ----------
+    testresults: List[TestResult]
+        The testresults.
+
+    Returns
+    -------
+    out : DataFrame
+        The testresults as data frame.
+    """
+    schema_test = T.StructType(
+        [
+            T.StructField("column", T.StringType(), True),
+            T.StructField("expression", T.StringType(), True),
+            T.StructField("id", T.StringType(), True),
+            T.StructField("metrics", T.ArrayType(T.StringType(), True), True),
+            T.StructField("source", T.StringType(), True),
+            T.StructField("title", T.StringType(), True),
+        ]
+    )
+
+    schema = T.StructType(
+        [
+            T.StructField("test", schema_test, True),
+            T.StructField("passed", T.BooleanType(), True),
+            T.StructField("skipped", T.BooleanType(), True),
+            T.StructField(
+                "values", T.MapType(T.StringType(), T.StringType()), True
+            ),
+            T.StructField("error", T.StringType(), True),
+            T.StructField(
+                "group_values", T.MapType(T.StringType(), T.StringType()), True
+            ),
+        ]
+    )
+    spark_session = SparkSession.builder.getOrCreate()
+    out = spark_session.createDataFrame(
+        test_results,
+        schema=schema,
+    )
+    return out
+
+
+def scan_errors_to_data_frame(scan_errors: list[ScanError]) -> DataFrame:
+    """
+    Convert ScanError to a data frame.
+
+    Parameters
+    ----------
+    scanerror: List[ScanError]
+        The scanerrors.
+
+    Returns
+    -------
+    out : DataFrame
+        The scanerrors as data frame.
+    """
+    schema = T.StructType(
+        [
+            T.StructField("message", T.StringType(), True),
+            T.StructField("exception", T.StringType(), True),
+        ]
+    )
+    spark_session = SparkSession.builder.getOrCreate()
+    out = spark_session.createDataFrame(scan_errors, schema=schema)
+    return out
+
+
+def convert_scan_result_to_spark_data_frames(
+    scan_result: ScanResult,
+) -> tuple[DataFrame, DataFrame, DataFrame]:
+    """
+    Convert the scan results to three Spark data frames.
+
+    Parameters
+    ------------
+    scan_result : ScanResult
+        The scan result.
+
+    Returns
+    --------
+    tuple[DataFrame, DataFrame, DataFrame] :
+        A Spark data frame with the:
+        1. measurements;
+        2. test results;
+        3. scan errors;
+    """
+    return (
+        measurements_to_data_frame(scan_result.measurements),
+        test_results_to_data_frame(scan_result.test_results),
+        scan_errors_to_data_frame(scan_result.errors),
+    )
+
+
 def execute(
     scan_definition: str | Path,
     df: DataFrame,
     *,
     soda_server_client: SodaServerClient | None = None,
+    as_frames: bool | None = False,
 ) -> ScanResult:
     """
     Execute a scan on a data frame.
@@ -290,6 +433,8 @@ def execute(
         The data frame to be scanned.
     soda_server_client : Optional[SodaServerClient] (default : None)
         A soda server client.
+    as_frames : bool (default : False)
+        Flag to return results in Dataframe
 
     Returns
     -------
@@ -298,6 +443,13 @@ def execute(
     """
     scan_yml = create_scan_yml(scan_definition)
     df.createOrReplaceTempView(scan_yml.table_name)
+
     scan = create_scan(scan_yml, soda_server_client=soda_server_client)
     scan.execute()
-    return scan.scan_result
+
+    if as_frames:
+        result = convert_scan_result_to_spark_data_frames(scan.scan_result)
+    else:
+        result = scan.scan_result
+
+    return result